requirements.txt
PyGithub
requests
gitpython
